Developed the structure a bit further
TriviaProvider was refactored to return an answer and a question as a String, and answer options as a string array

LabelHandler and ButtonHandler classes were introduced.

Added an enum file for the sake of presenting game captions.

Added logically working constraints, has to be fine-tuned and tested.

Pilot non-crashing version was assembled.